pyproject.toml
[tool.poetry]
name = "langchain-kr"
version = "0.1.1"
description = "Ultimate LangChain Tutorial written in Korean by TeddyNote."
authors = ["teddy <teddylee777@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# 언어 및 자연어 처리 관련 패키지
langchain = "0.*"
langchain-core = "0.*"
langchain-experimental = "0.*"
langchain-community = "0.*"
langchain-openai = "0.*"
langchain-anthropic = "0.*"
langchain-text-splitters = "0.*"
langchain-elasticsearch = "0.*"
langchain-chroma = "0.*"
langchain-cohere = "0.*"
langchain-milvus = "0.*"
langchain-google-genai = "2.*"
langchain-huggingface = "0.*"
langchain-azure-ai = "0.*"
langchain-teddynote = "0.*"
langchainhub = "0.*"
langgraph = "0.*"
langsmith = "0.*"
huggingface-hub = "0.*"
openai = "1.*"
deepl = "1.*"
kiwipiepy = "0.*"
konlpy = "0.*"

# 데이터 처리 및 분석 관련 패키지
pandas = "2.*"
rank-bm25 = "0.*"

# 데이터베이스 및 캐시 관련 패키지
redis = "5.*"
chromadb = "0.*"

# PDF 및 파일 처리 관련 패키지
pymupdf = "1.*"
pypdf = "4.*"
pdfplumber = "0.*"
pdfminer-six = "20231228"
pymupdf4llm = "0.*"

# 시각화 및 사용자 인터페이스 관련 패키지
matplotlib = "3.*"
streamlit = "1.*"
jupyter = "1.*"
notebook = "7.*"

# 딥러닝 및 머신러닝 관련 패키지
torch = "2.2.0"
torchvision = "0.*"
faiss-cpu = "1.*"
open-clip-torch = "2.*"

# 기타 유틸리티 및 필수 패키지
python-dotenv = "1.*"
pydantic = "2.*"
lxml = "5.*"
pillow = "10.*"
lark = "1.*"
ragas = "0.*"
unstructured = {version = "0.*", extras = ["all-docs"]}
arxiv = "2.*"
tiktoken = "0.*"
tenacity = "8.*"
pymilvus = "2.*"
google-search-results = "2.*"
protobuf = "3.*"
sqlalchemy = "2.*"
llama-index-core = "0.*"
llama-parse = "0.*"
llama-index-readers-file = "0.*"
flashrank = "0.*"
docx2txt = "0.*"
nest-asyncio = "1.*"
rapidocr-onnxruntime = "1.*"
seaborn = "0.*"
grandalf = "0.*"
rouge-score = "0.*"
langchain-ollama = "0.*"
mypy = "1.*"
pinecone = "5.*"
wikipedia = "1.*"
scikit-learn = "1.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
